Add files via upload
Here is the tutorial for handling NetCDF data (*.nc). At the beginning there is a link to a file on gitHub that explains how to get the data from IPCC DDC AR5. ... In the tutorial there is also a chapter about French Guiana and FORMIND, with which probably not everyone can directly begin something. But I left it in anyway, because it is a good example for 'make_ts'.R. make_ts.R' contains a function that turns the raw data of the *.nc files time series into time series for certain points of interest. A short documentation about the function can be found in the script.